rename package keystore to crypto
