+++ internal/check-imports/src/imports.ts
@@ -4,10 +4,10 @@ import * as address from '@fuel-ts/address';
 import * as contract from '@fuel-ts/contract';
 // forc-bin
 // fuels-gauge
+import * as crypto from '@fuel-ts/crypto';
 import * as hasher from '@fuel-ts/hasher';
 import * as hdwallet from '@fuel-ts/hdwallet';
 import * as interfaces from '@fuel-ts/interfaces';
-import * as keystore from '@fuel-ts/keystore';
 import * as math from '@fuel-ts/math';
 import * as merkle from '@fuel-ts/merkle';
 import * as mnemonic from '@fuel-ts/mnemonic';
@@ -36,7 +36,7 @@ log([
   hasher,
   hdwallet,
   interfaces,
-  keystore,
+  crypto,
   math,
   merkle,
   mnemonic,
