+++ internal/check-imports/src/references.ts
@@ -5,10 +5,10 @@ import { runTypegen } from '@fuel-ts/abi-typegen/runTypegen';
 import { Address } from '@fuel-ts/address';
 import { BaseAssetId } from '@fuel-ts/address/configs';
 import { ContractFactory } from '@fuel-ts/contract';
+import { encrypt, decrypt } from '@fuel-ts/crypto';
 import { hashTransaction, hashMessage } from '@fuel-ts/hasher';
 import { HDWallet } from '@fuel-ts/hdwallet';
 import { AbstractPredicate } from '@fuel-ts/interfaces';
-import { encrypt, decrypt } from '@fuel-ts/keystore';
 import { BN } from '@fuel-ts/math';
 import { DEFAULT_PRECISION, DEFAULT_MIN_PRECISION } from '@fuel-ts/math/configs';
 import { SparseMerkleTree, constructTree } from '@fuel-ts/merkle';
