+++ packages/abi-coder/package.json
@@ -29,7 +29,7 @@
     "@ethersproject/properties": "^5.7.0",
     "@ethersproject/sha2": "^5.7.0",
     "@ethersproject/strings": "^5.7.0",
-    "@fuel-ts/keystore": "workspace:*",
+    "@fuel-ts/crypto": "workspace:*",
     "@fuel-ts/math": "workspace:*",
     "@fuel-ts/versions": "workspace:*",
     "type-fest": "^3.1.0"
