+++ packages/mnemonic/package.json
@@ -26,7 +26,7 @@
     "@ethersproject/bytes": "^5.7.0",
     "@ethersproject/pbkdf2": "^5.7.0",
     "@ethersproject/sha2": "^5.7.0",
-    "@fuel-ts/keystore": "workspace:*",
+    "@fuel-ts/crypto": "workspace:*",
     "@fuel-ts/wordlists": "workspace:*"
   }
 }
