+++ packages/mnemonic/src/mnemonic.ts
@@ -3,7 +3,7 @@ import type { BytesLike } from '@ethersproject/bytes';
 import { hexDataSlice, concat, hexlify, arrayify } from '@ethersproject/bytes';
 import { pbkdf2 } from '@ethersproject/pbkdf2';
 import { computeHmac, sha256, SupportedAlgorithm } from '@ethersproject/sha2';
-import { randomBytes } from '@fuel-ts/keystore';
+import { randomBytes } from '@fuel-ts/crypto';
 import { english } from '@fuel-ts/wordlists';
 
 import type { MnemonicPhrase } from './utils';
